+++ src/rdmnet/discovery/avahi/disc_platform_defs.h
