+++ src/rdmnet/discovery/discovered_broker.h
@@ -1,0 +1,51 @@
+/******************************************************************************
+ * Copyright 2019 ETC Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ ******************************************************************************
+ * This file is a part of RDMnet. For more information, go to:
+ * https://github.com/ETCLabs/RDMnet
+ *****************************************************************************/
+
+#ifndef DISCOVERED_BROKER_H_
+#define DISCOVERED_BROKER_H_
+
+#include "rdmnet/core/discovery.h"
+#include "disc_platform_defs.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef struct DiscoveredBroker DiscoveredBroker;
+struct DiscoveredBroker
+{
+  char full_service_name[RDMNET_DISC_SERVICE_NAME_MAX_LENGTH];
+  RdmnetBrokerDiscInfo info;
+  RdmnetScopeMonitorRef* monitor_ref;
+  RdmnetDiscoveredBrokerPlatformData platform_data;
+  DiscoveredBroker* next;
+};
+
+DiscoveredBroker* discovered_broker_lookup_by_name(DiscoveredBroker* list_head, const char* full_name);
+DiscoveredBroker* discovered_broker_new(const char* service_name, const char* full_service_name);
+void discovered_broker_insert(DiscoveredBroker** list_head_ptr, DiscoveredBroker* new_db);
+void discovered_broker_remove(DiscoveredBroker** list_head_ptr, const DiscoveredBroker* db);
+void discovered_broker_delete(DiscoveredBroker* db);
+void discovered_broker_free_platform_resources(DiscoveredBroker* db);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* DISCOVERED_BROKER_H_ */
