+++ tests/CMakeLists.txt
@@ -1,6 +1,3 @@
 # Main test module. 
 
-message(STATUS "Adding GoogleTest dependency...")
-include(AddGoogleTest)
-
 add_subdirectory(unit)
