+++ tests/unit/discovery/common/CMakeLists.txt
@@ -2,13 +2,17 @@
 
 add_executable(test_discovery_common
   test_discovery_common.cpp
-  ${RDMNET_SRC}/rdmnet/discovery/common.h
-  ${RDMNET_SRC}/rdmnet/discovery/common.c
+#  ${RDMNET_SRC}/rdmnet/discovery/common.h
+#  ${RDMNET_SRC}/rdmnet/discovery/common.c
 
   ${RDMNET_INCLUDE}/rdmnet_mock/core.h
   ${RDMNET_SRC}/rdmnet/core/util.c
 )
-target_include_directories(test_discovery_common PRIVATE ${RDMNET_INCLUDE} ${RDMNET_SRC} ${CMAKE_CURRENT_LIST_DIR})
+target_include_directories(test_discovery_common PRIVATE
+  ${RDMNET_INCLUDE}
+  ${RDMNET_SRC}
+  ${CMAKE_CURRENT_LIST_DIR}
+)
 target_link_libraries(test_discovery_common PRIVATE EtcPalMock gtest_main meekrosoft::fff)
 set_target_properties(test_discovery_common PROPERTIES CXX_STANDARD 11 FOLDER tests)
 gtest_add_tests(TARGET test_discovery_common SKIP_DEPENDENCY)
